pyproject.toml
[project]
name = "niowase-police"
version = "0.1.0"
description = "Discord niowase detection bot"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "discord-py>=2.7.0",
    "pillow>=12.1.1",
    "requests>=2.32.5",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.2.0",
]
